backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pymongo==4.6.0
motor==3.3.2
pytesseract==0.3.10
opencv-python==4.8.1.78
pillow==10.1.0
pymupdf==1.23.8
qrcode[pil]==7.4.2
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
python-dotenv==1.0.0
pydantic==2.5.0
numpy==1.24.3